+++ examples/vue/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @tradecanvas/example-vue
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @tradecanvas/chart@0.5.0
+
 ## 0.0.3
 
 ### Patch Changes